rn: Flatten the styles folder
