--- packages/edge-login-ui-rn/src/styles/ScreenStyle.js
@@ -1,10 +1,0 @@
-// @flow
-
-import * as Colors from '../constants/Colors.js'
-
-export const ScreenStyle = {
-  flex: 1,
-  width: '100%',
-  height: '100%',
-  backgroundColor: Colors.WHITE
-}
