--- packages/edge-login-ui-rn/src/styles/common/Fonts.js
@@ -1,8 +1,0 @@
-// @flow
-
-const CreateAccountFont = {
-  defaultFontSize: 14,
-  headerFontSize: 17
-}
-
-export { CreateAccountFont }
