+++ packages/edge-login-ui-rn/src/styles/index.js
@@ -1,6 +1,18 @@
 // @flow
 
-export * from './common/ButtonStyles'
-export * from './common/Fonts'
-export * from './common/FormFieldStyle'
-export * from './ScreenStyle'
+import * as Colors from '../constants/Colors.js'
+
+export * from './ButtonStyles.js'
+export * from './FormFieldStyle.js'
+
+export const CreateAccountFont = {
+  defaultFontSize: 14,
+  headerFontSize: 17
+}
+
+export const ScreenStyle = {
+  flex: 1,
+  width: '100%',
+  height: '100%',
+  backgroundColor: Colors.WHITE
+}
